moved tokens tools to Token component
