+++ src/Hal/Component/Parser/CodeParser/ArgumentsParser.php
@@ -5,7 +5,7 @@ use Hal\Component\Parser\Exception\IncorrectSyntaxException;
 use Hal\Component\Reflected\Argument;
 use Hal\Component\Parser\Resolver\NamespaceResolver;
 use Hal\Component\Parser\Searcher;
-use Hal\Component\Parser\Token;
+use Hal\Component\Token\Token;
 
 
 class ArgumentsParser
