+++ src/Hal/Component/Parser/CodeParser/ClassParser.php
@@ -7,7 +7,7 @@ use Hal\Component\Reflected\KlassAnonymous;
 use Hal\Component\Reflected\KlassInterface;
 use Hal\Component\Parser\Resolver\NamespaceResolver;
 use Hal\Component\Parser\Searcher;
-use Hal\Component\Parser\Token;
+use Hal\Component\Token\Token;
 
 class ClassParser
 {
