+++ src/Hal/Component/Parser/Helper/TypeResolver.php
@@ -1,7 +1,7 @@
 <?php
 namespace Hal\Component\Parser\Helper;
 
-use Hal\Component\Parser\Token;
+use Hal\Component\Token\Token;
 
 class TypeResolver
 {
