+++ src/Hal/Component/Reflected/Method.php
@@ -3,7 +3,7 @@ namespace Hal\Component\Reflected;
 
 
 use Hal\Component\Parser\Helper\TypeResolver;
-use Hal\Component\Parser\Token;
+use Hal\Component\Token\Token;
 
 class Method
 {
