+++ src/Hal/Metrics/Complexity/Component/McCabe/McCabe.php
@@ -8,7 +8,7 @@
  */
 
 namespace Hal\Metrics\Complexity\Component\McCabe;
-use Hal\Component\Parser\Token;
+use Hal\Component\Token\Token;
 use Hal\Component\Token\TokenCollection;
 
 /**
