+++ tests/Hal/Component/Parser/Helper/NamespaceReplacerTest.php
@@ -2,8 +2,8 @@
 
 namespace Test;
 use Hal\Component\Parser\Helper\NamespaceReplacer;
-use Hal\Component\Parser\Token;
-use Hal\Component\Parser\Tokenizer;
+use Hal\Component\Token\Token;
+use Hal\Component\Token\Tokenizer;
 
 /**
  * @group parser
