+++ tests/Hal/Component/Parser/SubCase/AnonymousClassesTest.php
@@ -5,8 +5,8 @@ use Hal\Component\Parser\CodeParser;
 use Hal\Component\Reflected\File;
 use Hal\Component\Parser\Resolver\NamespaceResolver;
 use Hal\Component\Parser\Searcher;
-use Hal\Component\Parser\Token;
-use Hal\Component\Parser\Tokenizer;
+use Hal\Component\Token\Token;
+use Hal\Component\Token\Tokenizer;
 
 /**
  * @group parser
