+++ tests/Hal/Component/Tree/HashMapTest.php
@@ -1,8 +1,8 @@
 <?php
 
 namespace Test;
-use Hal\Component\Parser\Token;
-use Hal\Component\Parser\Tokenizer;
+use Hal\Component\Token\Token;
+use Hal\Component\Token\Tokenizer;
 use Hal\Component\Tree\HashMap;
 use Hal\Component\Tree\Node;
